Move data-modeling utils to metabase-lib (#25855)
