+++ frontend/src/metabase/query_builder/components/view/sidebars/DatasetManagementSection/DatasetMetadataStrengthIndicator/DatasetMetadataStrengthIndicator.tsx
@@ -1,13 +1,12 @@
 import React from "react";
 import { t } from "ttag";
 
-import ProgressBar from "metabase/components/ProgressBar";
 import Tooltip from "metabase/components/Tooltip";
 
 import { color } from "metabase/lib/colors";
-import { getDatasetMetadataCompletenessPercentage } from "metabase/lib/data-modeling/metadata";
 import { useHover } from "metabase/hooks/use-hover";
 import Question from "metabase-lib/lib/Question";
+import { getDatasetMetadataCompletenessPercentage } from "metabase-lib/lib/metadata/utils/models";
 
 import {
   Root,
