+++ frontend/src/metabase/query_builder/components/view/sidebars/ModelCacheManagementSection/ModelCacheManagementSection.tsx
@@ -4,11 +4,11 @@ import moment from "moment-timezone";
 import { connect } from "react-redux";
 
 import PersistedModels from "metabase/entities/persisted-models";
-import { checkCanRefreshModelCache } from "metabase/lib/data-modeling/utils";
 
 import { ModelCacheRefreshStatus } from "metabase-types/api";
 import Question from "metabase-lib/lib/Question";
 
+import { checkCanRefreshModelCache } from "metabase-lib/lib/metadata/utils/models";
 import {
   Row,
   StatusContainer,
