+++ frontend/src/metabase/query_builder/selectors.js
@@ -17,8 +17,6 @@ import { getComputedSettingsForSeries } from "metabase/visualizations/lib/settin
 import { getCardUiParameters } from "metabase/parameters/utils/cards";
 import { normalizeParameterValue } from "metabase/parameters/utils/parameter-values";
 
-import { isAdHocModelQuestion } from "metabase/lib/data-modeling/utils";
-
 import Databases from "metabase/entities/databases";
 import Timelines from "metabase/entities/timelines";
 
@@ -39,6 +37,7 @@ import { isPK } from "metabase-lib/lib/types/utils/isa";
 import Mode from "metabase-lib/lib/Mode";
 import NativeQuery from "metabase-lib/lib/queries/NativeQuery";
 import Question from "metabase-lib/lib/Question";
+import { isAdHocModelQuestion } from "metabase-lib/lib/metadata/utils/models";
 
 export const getUiControls = state => state.qb.uiControls;
 const getQueryStatus = state => state.qb.queryStatus;
