+++ frontend/src/metabase/visualizations/components/TableInteractive/TableInteractive.jsx
@@ -24,7 +24,6 @@ import {
   isColumnRightAligned,
 } from "metabase/visualizations/lib/table";
 import { getColumnExtent } from "metabase/visualizations/lib/utils";
-import { isAdHocModelQuestionCard } from "metabase/lib/data-modeling/utils";
 import { getScrollBarSize } from "metabase/lib/dom";
 import { zoomInRow } from "metabase/query_builder/actions";
 import { getQueryBuilderMode } from "metabase/query_builder/selectors";
@@ -37,6 +36,7 @@ import { isID, isPK, isFK } from "metabase-lib/lib/types/utils/isa";
 import { fieldRefForColumn } from "metabase-lib/lib/queries/utils/dataset";
 import Dimension from "metabase-lib/lib/Dimension";
 import { memoizeClass } from "metabase-lib/lib/utils";
+import { isAdHocModelQuestionCard } from "metabase-lib/lib/metadata/utils/models";
 import MiniBar from "../MiniBar";
 import {
   ExpandButton,
